feat(utils): added getBase64 to the utils file
