+++ src/components/sections/Projects/Projects.tsx
@@ -1,6 +1,6 @@
 import { Container, Title } from '@/components/ui';
 import { projects } from '@/src/data/data';
-import { getBase64 } from '@/src/lib/getBase64';
+import { getBase64 } from '@/src/lib/utils';
 import { ProjectCards } from './components/ProjectCards';
 
 export const Projects = async () => {
@@ -9,12 +9,12 @@ export const Projects = async () => {
 	const blurredImages = await Promise.all(images.map((image) => getBase64(image)));
 
 	return (
-		<section className="scroll-mt-[80px] pb-[100px] md:scroll-mt-[120px]" id="projects">
-			<Container className="flex flex-col">
-				<Title className="text-center text-[24px] font-bold md:text-[48px]" tag="h2">
+		<section className='scroll-mt-[80px] pb-[100px] md:scroll-mt-[120px]' id='projects'>
+			<Container className='flex flex-col'>
+				<Title className='text-center text-[24px] font-bold md:text-[48px]' tag='h2'>
 					Projects
 				</Title>
-				<div className="mt-[10px] text-center text-[18px] text-gray md:mt-[30px] md:text-[32px]">
+				<div className='mt-[10px] text-center text-[18px] text-gray md:mt-[30px] md:text-[32px]'>
 					Things I’ve built so far
 				</div>
 				<ProjectCards blurredImages={blurredImages} projects={revercedProjects} />
